+++ util/src/main/java/com/google/cloud/hadoop/util/AsyncWriteChannelOptions.java
@@ -29,6 +29,9 @@ public abstract class AsyncWriteChannelOptions {
   /** Default upload chunk size. */
   public static final int UPLOAD_CHUNK_SIZE_DEFAULT = 64 * 1024 * 1024;
 
+  /** Default upload cache size. */
+  public static final int UPLOAD_CACHE_SIZE_DEFAULT = 0;
+
   /** Default of whether to use direct upload. */
   public static final boolean DIRECT_UPLOAD_ENABLED_DEFAULT = false;
 
@@ -47,6 +50,7 @@ public abstract class AsyncWriteChannelOptions {
         .setBufferSize(BUFFER_SIZE_DEFAULT)
         .setPipeBufferSize(PIPE_BUFFER_SIZE_DEFAULT)
         .setUploadChunkSize(UPLOAD_CHUNK_SIZE_DEFAULT)
+        .setUploadCacheSize(UPLOAD_CACHE_SIZE_DEFAULT)
         .setDirectUploadEnabled(DIRECT_UPLOAD_ENABLED_DEFAULT);
   }
 
@@ -55,6 +59,8 @@ public abstract class AsyncWriteChannelOptions {
   public abstract int getPipeBufferSize();
 
   public abstract int getUploadChunkSize();
+
+  public abstract int getUploadCacheSize();
 
   public abstract boolean isDirectUploadEnabled();
 
@@ -67,6 +73,8 @@ public abstract class AsyncWriteChannelOptions {
     public abstract Builder setPipeBufferSize(int pipeBufferSize);
 
     public abstract Builder setUploadChunkSize(int uploadChunkSize);
+
+    public abstract Builder setUploadCacheSize(int uploadCacheSize);
 
     public abstract Builder setDirectUploadEnabled(boolean directUploadEnabled);